--- a/test/integration/apiserver/flowcontrol/fs_condition_test.go
+++ b/test/integration/apiserver/flowcontrol/fs_condition_test.go
@@ -17,7 +17,6 @@ limitations under the License.
 package flowcontrol
 
 import (
-	"context"
 	"encoding/json"
 	"testing"
 	"time"
@@ -38,14 +37,10 @@ import (
 func TestConditionIsolation(t *testing.T) {
 	defer featuregatetesting.SetFeatureGateDuringTest(t, utilfeature.DefaultFeatureGate, genericfeatures.APIPriorityAndFairness, true)()
 	// NOTE: disabling the feature should fail the test
-	kubeConfig, closeFn := setup(t, 10, 10)
+	ctx, kubeConfig, closeFn := setup(t, 10, 10)
 	defer closeFn()
 
 	loopbackClient := clientset.NewForConfigOrDie(kubeConfig)
-
-	stopCh := make(chan struct{})
-	defer close(stopCh)
-	ctx := context.Background()
 
 	fsOrig := fcboot.SuggestedFlowSchemas[0]
 	t.Logf("Testing Status Condition isolation in FlowSchema %q", fsOrig.Name)
@@ -60,7 +55,7 @@ func TestConditionIsolation(t *testing.T) {
 		}
 		dangleOrig = getCondition(fsGot.Status.Conditions, flowcontrol.FlowSchemaConditionDangling)
 		return dangleOrig != nil, nil
-	}, stopCh)
+	}, ctx.Done())
 
 	ssaType := flowcontrol.FlowSchemaConditionType("test-ssa")
 	patchSSA := flowcontrolapply.FlowSchema(fsOrig.Name).