--- a/test/integration/apiserver/max_request_body_bytes_test.go
+++ b/test/integration/apiserver/max_request_body_bytes_test.go
@@ -26,11 +26,16 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/kubernetes/test/integration/framework"
+	"k8s.io/kubernetes/test/utils/ktesting"
 )
 
 // Tests that the apiserver limits the resource size in write operations.
 func TestMaxResourceSize(t *testing.T) {
-	clientSet, _, tearDownFn := framework.StartTestServer(t, framework.TestServerSetup{})
+	_, ctx := ktesting.NewTestContext(t)
+	ctx, cancel := context.WithCancel(ctx)
+	defer cancel()
+
+	clientSet, _, tearDownFn := framework.StartTestServer(ctx, t, framework.TestServerSetup{})
 	defer tearDownFn()
 
 	hugeData := []byte(strings.Repeat("x", 3*1024*1024+1))
@@ -40,7 +45,7 @@ func TestMaxResourceSize(t *testing.T) {
 	c := clientSet.CoreV1().RESTClient()
 	t.Run("Create should limit the request body size", func(t *testing.T) {
 		err := c.Post().AbsPath("/api/v1/namespaces/default/pods").
-			Body(hugeData).Do(context.TODO()).Error()
+			Body(hugeData).Do(ctx).Error()
 		if err == nil {
 			t.Fatalf("unexpected no error")
 		}
@@ -56,14 +61,14 @@ func TestMaxResourceSize(t *testing.T) {
 			Name: "test",
 		},
 	}
-	_, err := clientSet.CoreV1().Secrets("default").Create(context.TODO(), secret, metav1.CreateOptions{})
+	_, err := clientSet.CoreV1().Secrets("default").Create(ctx, secret, metav1.CreateOptions{})
 	if err != nil {
 		t.Fatal(err)
 	}
 
 	t.Run("Update should limit the request body size", func(t *testing.T) {
 		err = c.Put().AbsPath("/api/v1/namespaces/default/secrets/test").
-			Body(hugeData).Do(context.TODO()).Error()
+			Body(hugeData).Do(ctx).Error()
 		if err == nil {
 			t.Fatalf("unexpected no error")
 		}
@@ -74,7 +79,7 @@ func TestMaxResourceSize(t *testing.T) {
 	})
 	t.Run("Patch should limit the request body size", func(t *testing.T) {
 		err = c.Patch(types.JSONPatchType).AbsPath("/api/v1/namespaces/default/secrets/test").
-			Body(hugeData).Do(context.TODO()).Error()
+			Body(hugeData).Do(ctx).Error()
 		if err == nil {
 			t.Fatalf("unexpected no error")
 		}
@@ -89,7 +94,7 @@ func TestMaxResourceSize(t *testing.T) {
 		}
 		patchBody := []byte(`[{"op":"add","path":"/foo","value":` + strings.Repeat("[", 3*1024*1024/2-100) + strings.Repeat("]", 3*1024*1024/2-100) + `}]`)
 		err = rest.Patch(types.JSONPatchType).AbsPath("/api/v1/namespaces/default/secrets/test").
-			Body(patchBody).Do(context.TODO()).Error()
+			Body(patchBody).Do(ctx).Error()
 		if err != nil && !apierrors.IsBadRequest(err) {
 			t.Errorf("expected success or bad request err, got %v", err)
 		}
@@ -100,7 +105,7 @@ func TestMaxResourceSize(t *testing.T) {
 		}
 		patchBody := []byte(`[{"op":"add","path":"/foo","value":0` + strings.Repeat(" ", 3*1024*1024-100) + `}]`)
 		err = rest.Patch(types.JSONPatchType).AbsPath("/api/v1/namespaces/default/secrets/test").
-			Body(patchBody).Do(context.TODO()).Error()
+			Body(patchBody).Do(ctx).Error()
 		if err != nil {
 			t.Errorf("unexpected error: %v", err)
 		}
@@ -111,7 +116,7 @@ func TestMaxResourceSize(t *testing.T) {
 		}
 		patchBody := []byte(`{"value":` + strings.Repeat("[", 3*1024*1024/2-100) + strings.Repeat("]", 3*1024*1024/2-100) + `}`)
 		err = rest.Patch(types.MergePatchType).AbsPath("/api/v1/namespaces/default/secrets/test").
-			Body(patchBody).Do(context.TODO()).Error()
+			Body(patchBody).Do(ctx).Error()
 		if err != nil && !apierrors.IsBadRequest(err) {
 			t.Errorf("expected success or bad request err, got %v", err)
 		}
@@ -122,7 +127,7 @@ func TestMaxResourceSize(t *testing.T) {
 		}
 		patchBody := []byte(`{"value":0` + strings.Repeat(" ", 3*1024*1024-100) + `}`)
 		err = rest.Patch(types.MergePatchType).AbsPath("/api/v1/namespaces/default/secrets/test").
-			Body(patchBody).Do(context.TODO()).Error()
+			Body(patchBody).Do(ctx).Error()
 		if err != nil {
 			t.Errorf("unexpected error: %v", err)
 		}
@@ -133,7 +138,7 @@ func TestMaxResourceSize(t *testing.T) {
 		}
 		patchBody := []byte(`{"value":` + strings.Repeat("[", 3*1024*1024/2-100) + strings.Repeat("]", 3*1024*1024/2-100) + `}`)
 		err = rest.Patch(types.StrategicMergePatchType).AbsPath("/api/v1/namespaces/default/secrets/test").
-			Body(patchBody).Do(context.TODO()).Error()
+			Body(patchBody).Do(ctx).Error()
 		if err != nil && !apierrors.IsBadRequest(err) {
 			t.Errorf("expected success or bad request err, got %v", err)
 		}
@@ -144,7 +149,7 @@ func TestMaxResourceSize(t *testing.T) {
 		}
 		patchBody := []byte(`{"value":0` + strings.Repeat(" ", 3*1024*1024-100) + `}`)
 		err = rest.Patch(types.StrategicMergePatchType).AbsPath("/api/v1/namespaces/default/secrets/test").
-			Body(patchBody).Do(context.TODO()).Error()
+			Body(patchBody).Do(ctx).Error()
 		if err != nil {
 			t.Errorf("unexpected error: %v", err)
 		}
@@ -155,7 +160,7 @@ func TestMaxResourceSize(t *testing.T) {
 		}
 		patchBody := []byte(`{"value":` + strings.Repeat("[", 3*1024*1024/2-100) + strings.Repeat("]", 3*1024*1024/2-100) + `}`)
 		err = rest.Patch(types.ApplyPatchType).Param("fieldManager", "test").AbsPath("/api/v1/namespaces/default/secrets/test").
-			Body(patchBody).Do(context.TODO()).Error()
+			Body(patchBody).Do(ctx).Error()
 		if err != nil && !apierrors.IsBadRequest(err) {
 			t.Errorf("expected success or bad request err, got %#v", err)
 		}
@@ -166,14 +171,14 @@ func TestMaxResourceSize(t *testing.T) {
 		}
 		patchBody := []byte(`{"apiVersion":"v1","kind":"Secret"` + strings.Repeat(" ", 3*1024*1024-100) + `}`)
 		err = rest.Patch(types.ApplyPatchType).Param("fieldManager", "test").AbsPath("/api/v1/namespaces/default/secrets/test").
-			Body(patchBody).Do(context.TODO()).Error()
+			Body(patchBody).Do(ctx).Error()
 		if err != nil {
 			t.Errorf("unexpected error: %v", err)
 		}
 	})
 	t.Run("Delete should limit the request body size", func(t *testing.T) {
 		err = c.Delete().AbsPath("/api/v1/namespaces/default/secrets/test").
-			Body(hugeData).Do(context.TODO()).Error()
+			Body(hugeData).Do(ctx).Error()
 		if err == nil {
 			t.Fatalf("unexpected no error")
 		}
@@ -197,7 +202,7 @@ values: ` + strings.Repeat("[", 3*1024*1024))
 			SetHeader("Content-Type", "application/yaml").
 			AbsPath("/api/v1/namespaces/default/configmaps").
 			Body(yamlBody).
-			DoRaw(context.TODO())
+			DoRaw(ctx)
 		if !apierrors.IsRequestEntityTooLargeError(err) {
 			t.Errorf("expected too large error, got %v", err)
 		}
@@ -220,7 +225,7 @@ values: ` + strings.Repeat("[", 3*1024*1024/2-500) + strings.Repeat("]", 3*1024*
 			SetHeader("Content-Type", "application/yaml").
 			AbsPath("/api/v1/namespaces/default/configmaps").
 			Body(yamlBody).
-			DoRaw(context.TODO())
+			DoRaw(ctx)
 		if !apierrors.IsBadRequest(err) {
 			t.Errorf("expected bad request, got %v", err)
 		}
@@ -243,7 +248,7 @@ values: ` + strings.Repeat("[", 3*1024*1024-1000))
 			SetHeader("Content-Type", "application/yaml").
 			AbsPath("/api/v1/namespaces/default/configmaps").
 			Body(yamlBody).
-			DoRaw(context.TODO())
+			DoRaw(ctx)
 		if !apierrors.IsBadRequest(err) {
 			t.Errorf("expected bad request, got %v", err)
 		}
@@ -264,7 +269,7 @@ values: ` + strings.Repeat("[", 3*1024*1024-1000))
 			SetHeader("Content-Type", "application/json").
 			AbsPath("/api/v1/namespaces/default/configmaps").
 			Body(jsonBody).
-			DoRaw(context.TODO())
+			DoRaw(ctx)
 		if !apierrors.IsRequestEntityTooLargeError(err) {
 			t.Errorf("expected too large error, got %v", err)
 		}
@@ -288,7 +293,7 @@ values: ` + strings.Repeat("[", 3*1024*1024-1000))
 			SetHeader("Content-Type", "application/json").
 			AbsPath("/api/v1/namespaces/default/configmaps").
 			Body(jsonBody).
-			DoRaw(context.TODO())
+			DoRaw(ctx)
 		// TODO(liggitt): expect bad request on deep nesting, rather than success on dropped unknown field data
 		if err != nil && !apierrors.IsBadRequest(err) {
 			t.Errorf("expected bad request, got %v", err)
@@ -313,7 +318,7 @@ values: ` + strings.Repeat("[", 3*1024*1024-1000))
 			SetHeader("Content-Type", "application/json").
 			AbsPath("/api/v1/namespaces/default/configmaps").
 			Body(jsonBody).
-			DoRaw(context.TODO())
+			DoRaw(ctx)
 		if !apierrors.IsBadRequest(err) {
 			t.Errorf("expected bad request, got %v", err)
 		}