<!--  Thanks for sending a pull request!  Here are some tips for you:

1. If this is your first time, please read our contributor guidelines: https://git.k8s.io/community/contributors/guide/first-contribution.md#your-first-contribution and developer guide https://git.k8s.io/community/contributors/devel/development.md#development-guide
2. Please label this pull request according to what type of issue you are addressing, especially if this is a release targeted pull request. For reference on required PR/issue labels, read here:
https://git.k8s.io/community/contributors/devel/sig-release/release.md#issuepr-kind-label
3. Ensure you have added or ran the appropriate tests for your PR: https://git.k8s.io/community/contributors/devel/sig-testing/testing.md
4. If you want *faster* PR reviews, read how: https://git.k8s.io/community/contributors/guide/pull-requests.md#best-practices-for-faster-reviews
5. If the PR is unfinished, see how to mark it: https://git.k8s.io/community/contributors/guide/pull-requests.md#marking-unfinished-pull-requests
-->

#### What type of PR is this?
/kind feature
/sig scheduling

<!--
Add one of the following kinds:
/kind bug
/kind cleanup
/kind documentation
/kind feature

Optionally add one or more of the following kinds if applicable:
/kind api-change
/kind deprecation
/kind failing-test
/kind flake
/kind regression
-->

#### What this PR does / why we need it:

In part of the scheduler integration tests, for each test loop, we'll initialize the api server and scheduler, and tear them down after test finishes. Considering api server is quite time consuming to setup, I made some change:
1. Initialize api server for only once at the beginning of the test
2. Initialize scheduler every time in each test loop and tear it down after test

Take the `TestPostFilterPlugin` for example, before it costs ~ 53.273s, after, it costs ~ 14.079s.

- To achieve 1, I made some changes to the initialization util functions, see 4bec44c515a0a525ec7f31387aa4eb3af1ead1d8
- To achieve 2, some tests will setup both the apiserver and scheduler in the beginning, and call the reset() function manually after each test, together with deleting the pods created before, these are something I treat as micro-management, so 
    - I removed such logics, instead, I'll recreate the scheduler each time
    - Also removed the `reset()` functions, no need to care about the teardowns.
    - As a result, the test will slow for ~1s then, for we'll recreate the scheduler each time(only for those who create apiserver and scheduler at the beginning of the whole test)


Results:
```
before: ok k8s.io/kubernetes/test/integration/scheduler/plugins 320.507s
after: ok k8s.io/kubernetes/test/integration/scheduler/plugins 229.595s
``` 

#### Which issue(s) this PR fixes:
<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #

#### Special notes for your reviewer:

#### Does this PR introduce a user-facing change?
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below. If the PR requires additional action from users switching to the new release, include the string "action required".

For more information on release notes see: https://git.k8s.io/community/contributors/guide/release-notes.md
-->
```release-note
NONE
```

#### Additional documentation e.g., KEPs (Kubernetes Enhancement Proposals), usage docs, etc.:

<!--
This section can be blank if this pull request does not require a release note.

When adding links which point to resources within git repositories, like
KEPs or supporting documentation, please reference a specific commit and avoid
linking directly to the master branch. This ensures that links reference a
specific point in time, rather than a document that may change over time.

See here for guidance on getting permanent links to files: https://help.github.com/en/articles/getting-permanent-links-to-files

Please use the following format for linking documentation:
- [KEP]: <link>
- [Usage]: <link>
- [Other doc]: <link>
-->
```docs

```