@alculquicondor 2023-03-15
In some tests, we are removing the apiserver recreation for every case. In other tests, we are adding the scheduler recreation in every case.

Can you share what are the running times for the entire package before and after the change?