@kerthcet 2023-03-16
> 

after: ok  	k8s.io/kubernetes/test/integration/scheduler/plugins	229.595s

before: ok  	k8s.io/kubernetes/test/integration/scheduler/plugins	320.507s