@alculquicondor 2023-03-22
/hold
in case the command above fails on me