@kerthcet 2023-03-30
> Thinking aloud here... should all of these commits have been squashed into one?
> 
> I went through them one-by-one because I expected a large PR like this to be broken down into smaller, self-contained commit. But then I realized that I was re-reviewing old commits that got fixed in later commits, so I had to do a lot of back-and-forth to check whether a comment that I was about to make still applied.
> 
> But that's just a side note. Overall the change looks good, just one nit about documentation.

em.. what I learned from the community is not squash for better reviewing, or you can't know what changes later ... And you can organize the commits in the final. Do you think this not works for you?