@pohly 2023-03-30
I tend to prefer clean commits (can be merged right away when they are ready, can start from scratch when I need to, for example after coming back to a PR after a while) and use the GitHub features (changes since last review, force-push diff) to track what gets changed overall. Perhaps that's just me :shrug: