@alculquicondor 2023-03-30
Maybe you can split this PR in 2: one for the framework changes (plus one test update) and one for the rest of the test updates.

I think the way you structured the commits might allow to do this somewhat easily.