@pohly 2023-03-30
I'm fine with squashing into reasonable commits. My only remaining comment is about documenting `initTestSchedulerForFrameworkTest` which will be easy to check.