@kerthcet 2023-03-31
> This is the shutdown function returned by initTestSchedulerForFrameworkTest, right? That looks like a better approach than the stand-alone cleanupTestSchedulerForFrameworkTest function here.

I can't reply on the comment https://github.com/kubernetes/kubernetes/pull/116634#discussion_r1152971097, but the function is already removed, so let it go.