@pohly 2023-03-31
The first three commits are not clean yet: the first one mixes functional changes with dead code removal, and the last one fixes something in an earlier commit. I think you can simply squash all of them into one. The last commit is self-contained.