@kerthcet 2023-04-03
The `pull-kubernetes-verify-strict-lint` requires a lot of code changes.